`--clean` now cleans the output dir instead of deleting it
The most common behavior for a CLI `--clean` flag is to delete just the contents of the artifact / output folder, not the folder itself.

This ensures that the folder (and especially any required permissions or configuration) remains in place for subsequent runs, avoiding issues where the folder needs to be recreated.

Deleting only the contents is safer and more user-friendly, especially if other processes or scripts expect the folder to exist.

This commit closes https://github.com/ergogen/ergogen/pull/131