Add SVG generation option to ergogen processing
This change introduces a `--svg` or `--generate-svg` CLI flag, disabled by default, to generate SVG files for each outline independent of `debug`. This is necessary for the Ergogen GUI not to run in debug mode all the time, as well as for more control on the user side of what exports are included.

The flag is set to `false` by default not to modify CLI behavior if not supplied.